feat: 게시글 등록 / 수정 기능 구현 (#90)
* feat: 게시글 수정 구현 - 1

* feat: 게시글 등록 / 수정 기능 구현

+++ server/src/main/java/com/lookatme/server/board/dto/SimpleBoardResponseDto.java
@@ -16,7 +16,7 @@ import java.util.List;
 @AllArgsConstructor
 public class SimpleBoardResponseDto {
 
-    private int boardId;
+    private long boardId;
 
     private String userImage;
